requirements.txt
# Python dependencies for main.py
requests>=2.31.0
colorama>=0.4.6
psutil>=5.9.5
bcrypt>=4.1.3
psycopg2-binary>=2.9.9
PyYAML>=6.0.2
playwright>=1.47.0
pytest>=8.3.2